Move code from Client to Model
