+++ src/Pyz/Zed/Training/TrainingDependencyProvider.php
@@ -2,7 +2,6 @@
 
 namespace Pyz\Zed\Training;
 
-use Pyz\Zed\Training\Dependency\Facade\TrainingToEventBridge;
 use Spryker\Zed\Kernel\AbstractBundleDependencyProvider;
 use Spryker\Zed\Kernel\Container;
 
@@ -32,7 +31,7 @@ class TrainingDependencyProvider extends AbstractBundleDependencyProvider
     private function addEventFacade(Container $container): Container
     {
         $container[static::FACADE_EVENT] = function (Container $container) {
-            return new TrainingToEventBridge($container->getLocator()->event()->facade());
+            return $container->getLocator()->event()->facade();
         };
 
         return $container;
